Fix/preference chain (#5)
* wip reworked preference chain

* feat (preferance_chain): clustering logic for grouping Action through the after field.

The logic can be stated as : we want to group action as simple chain, linked by uniquely the `after` property, and that have the same topological structure. It means that the external dependencies cannont grow, and the node head of a chain cannot have multiple referenced which are not all part of the chain

* fix (preferance_chain): fix some logging and minor errors

* feat (preferance_chain): compute all possible resolution pathes

* refactor (preference_chain): update ChainLink structure and improve path computation

This commit refactors the ChainLink class by replacing the 'queue' attribute with a 'tail' attribute, enhancing the clarity of the chain's structure. The update also modifies the logic in the update_chains and merge_chains functions to accommodate this change.

Additionally, the compute_all_resolution_pathes function has been reworked to utilize permutations for generating valid paths, ensuring that head dependencies and external dependencies are properly validated. This improves the accuracy of the path computation process.

* wip: test preference_chain

* fix: use the new preference_chain in the cli